Updating MEAN stack tutorial files

+++ MEAN Stack Guide - Angular and Node/mean-course/backend/routes/posts.js
@@ -33,7 +33,8 @@ router.post("", checkAuth, multer({storage: storage}).single("image"), (req, res
     const post = new Post({
         title: req.body.title,
         content: req.body.content,
-        imagePath: url + "/images/" + req.file.filename
+        imagePath: url + "/images/" + req.file.filename,
+        creator: req.userData.userId
     });
     post.save().then(result => {
         res.status(201).json({
@@ -42,6 +43,11 @@ router.post("", checkAuth, multer({storage: storage}).single("image"), (req, res
                 ...result,
                 id: result._id,
             }
+        });
+    })
+    .catch(error => {
+        res.status(500).json({
+            message: "Creating a post failed."
         });
     });
 });
@@ -56,13 +62,21 @@ router.put("/:id", checkAuth, multer({storage: storage}).single("image"), (req, 
         _id: req.body.id,
         title: req.body.title,
         content: req.body.content,
-        imagePath: imagePath
+        imagePath: imagePath,
+        creator: req.userData.userId
     });
-    console.log(post);
-    Post.updateOne({_id: req.params.id}, post)
+    Post.updateOne({ _id: req.params.id, creator: req.userData.userId }, post)
     .then(result => {
-        console.log(result);
-        res.status(200).json({ message: "Update successful!" });
+        if (result.modifiedCount > 0) {
+            res.status(200).json({ message: "Update successful!" });
+        } else {
+            res.status(401).json({ message: "Not authorized." });
+        }
+    })
+    .catch(error => {
+        res.status(500).json({
+            message: "Couldn't update post."
+        });
     });
 });
 
@@ -86,6 +100,11 @@ router.get('', (req, res, next) => {
             posts: fetchedPosts,
             maxPosts: count
         });
+    })
+    .catch(error => {
+        res.status(500).json({
+            message: "Fetching posts failed."
+        });
     });
 });
 
@@ -96,13 +115,25 @@ router.get("/:id", (req, res, next) => {
         } else {
             res.status(404).json({message: 'Post not found!'});
         }
-    })
+    }).catch(error => {
+        res.status(500).json({
+            message: "Fetching post failed."
+        });
+    });
 });
 
 router.delete("/:id", checkAuth, (req, res, next) => {
-    Post.deleteOne({_id: req.params.id}).then(result => {
-        console.log(result);
-        res.status(200).json({ message: "Post deleted!" });
+    Post.deleteOne({ _id: req.params.id, creator: req.userData.userId }).then(result => {
+        if (result.deletedCount > 0) {
+            res.status(200).json({ message: "Deletion successful!" });
+        } else {
+            res.status(401).json({ message: "Not authorized." });
+        }
+    })
+    .catch(error => {
+        res.status(500).json({
+            message: "Deleting posts failed."
+        });
     });
 });
